WebSockets/tests/submissions/Opera/interfaces/WebSocket/bufferedAmount/007.html

<!doctype html>
<title>WebSockets: bufferedAmount for blob</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+':80/ws/echo');
  var data = "";
  var datasize = 10;  
  ws.onopen = function(e) {
   ws.binaryType = "blob";
   data = new ArrayBuffer(datasize);
   ws.send(data);
   assert_equals(data.size(), ws.bufferedAmount);
  }
  ws.onmessage = function(e) {
    assertInstanceOf(e.data,Blob);
    assert_equals(e.data.size, datasize);
    end();
  }
  
});
</script>
